Changed logbook page name to Plan Hike as well as history page to logbook. Added view of upcomming hikes and allowed to edit start time
